requirements.txt
streamlit
pandas
requests
langchain
langchain-community
ollama
newspaper3k